+++ spec/dummy/client/app/ror-auto-load-components/ConsoleLogsInAsyncServer.client.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import ConsoleLogsInAsyncServer from '../components/ConsoleLogsInAsyncServer';
+
+const ConsoleLogsInAsyncServerClient = ({ requestId }) => <ConsoleLogsInAsyncServer requestId={requestId} />;
+
+export default ConsoleLogsInAsyncServerClient;