+++ CHANGELOG.md
@@ -14,6 +14,12 @@ You can find the **package** version numbers from this repo's tags and below in 
 ## [Unreleased]
 *Add changes in master not yet tagged.*
 
+### Added
+- Added support for replaying console logs from asynchronous operations:
+  - New `replayServerAsyncOperationLogs` configuration option to enable/disable this feature
+  - When enabled, captures and replays console output from async operations during server-side rendering
+  - [PR 440](https://github.com/shakacode/react_on_rails_pro/pull/440) by [AbanoubGhadban](https://github.com/AbanoubGhadban)
+
 ### Removed
 - Drop support for EOL'd Ruby 2.7 [PR 365](https://github.com/shakacode/react_on_rails_pro/pull/365) by [ahangarha](https://github.com/ahangarha).
 - Drop support for React on Rails below 14.0.4 [PR 415](https://github.com/shakacode/react_on_rails_pro/pull/415) by [rameziophobia](https://github.com/rameziophobia).